Refact.Types

Documentation

data SrcSpan

A generic SrcSpan, usually this is converted immediately to a native representation. (For example a GHC SrcSpan or a HSE SrcSpan)

Constructors

SrcSpan

Fields

startLine :: !Int

startCol :: !Int

endLine :: !Int

endCol :: !Int

data RType

Types of expressions which we are able to replace.

Constructors

Expr

Decl

Type

Pattern

Stmt

ModuleName

Bind

Match

Import

data Refactoring a

Supported refactorings

Constructors

Replace

Fields

rtype :: RType

Type of expression to be replaced

pos :: a

Expression to replace

subts :: [(String, a)]

Substitutions to make

orig :: String

Replacement template

ModifyComment

Fields

pos :: a

Expression to replace

newComment :: String

InsertComment

Fields

pos :: a

Expression to replace

newComment :: String

Delete

Fields

rtype :: RType

Type of expression to be replaced

pos :: a

Expression to replace

RemoveAsKeyword
Takes the position of a import decl and removes the as keyword

Fields

pos :: a

Expression to replace